requirements.txt
# SOC Defense System Requirements
# Install with: pip install -r requirements.txt

# Core GUI
customtkinter>=5.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Elasticsearch
elasticsearch>=8.0.0
aiohttp>=3.8.0

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0
folium>=0.14.0

# Microsoft 365 Email Integration
msal>=1.24.0
requests>=2.31.0

# PDF Report Generation
reportlab>=4.0.0
WeasyPrint>=60.0  # Optional: for advanced PDF generation

# GeoIP (Optional)
geoip2>=4.7.0

# Other Utilities
python-dateutil>=2.8.0
pytz>=2023.3
